requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pycryptodome==3.20.0
requests==2.31.0
gunicorn
